requirements.txt
absl-py==2.1.0
alabaster==0.7.12
altgraph==0.17.2
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
auto-py-to-exe==2.46.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.9.0
blis==0.7.7
boto3==1.38.5
botocore==1.38.5
bottle==0.13.3
bottle-websocket==0.2.9
branca==0.7.2
bs4==0.0.2
cachetools==5.3.3
catalogue==2.0.7
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.4
comtypes==1.1.11
cryptography==42.0.5
cssselect==1.2.0
customtkinter==4.3.0
cycler==0.11.0
cymem==2.0.6
darkdetect==0.6.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docopt==0.6.2
docutils==0.18.1
Eel==0.18.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.2.0/en_core_web_md-3.2.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
fake-useragent==1.5.1
fastjsonschema==2.16.2
feedparser==6.0.9
filelock==3.15.4
Flask==3.1.0
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.28.5
fsspec==2024.9.0
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.5.4
geocoder==1.38.1
gevent==25.4.2
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.42
gnewsclient==1.12
google-ai-generativelanguage==0.6.15
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-texttospeech==2.26.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.2.1
grpcio==1.64.1
grpcio-status==1.62.3
gTTS==2.2.4
h5py==3.11.0
httplib2==0.22.0
idna==3.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
imagesize==1.3.0
importlib_metadata==7.1.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
it-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.2.0/it_core_news_sm-3.2.0-py3-none-any.whl
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.1.0
Js2Py==0.71
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.2
keras==3.3.3
keras-models==0.0.7
keyboard==0.13.5
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.3.2
lab-utils==0.5.12
langcodes==3.3.0
libclang==18.1.1
lxml==4.9.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.6
namex==0.0.8
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.1
nest-asyncio==1.5.5
networkx==3.3
nltk==3.7
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.5.5.64
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging==25.0
pandas==1.3.5
pandocfilters==1.5.0
parse==1.20.1
parso==0.8.3
pathy==0.6.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.0.0
pipwin==0.5.2
playsound==1.2.2
playsound2==0.1
plotly==5.10.0
plyfile==0.7.4
pocketsphinx==5.0.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.31
proto-plus==1.26.1
protobuf==4.25.3
psutil==5.9.2
psycopg2==2.9.5
pure-eval==0.2.2
pvporcupine==2.1.4
pvrecorder==1.1.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio @ file:///C:/Users/edoar/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pycparser==2.21
pydantic==1.8.2
PyDrive==1.3.1
pyee==11.1.0
pygame==2.1.2
PyGithub==2.2.0
Pygments==2.18.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyjsparser==2.7.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyodide-py==0.25.1
pyparsing==3.0.6
pyperclip==1.8.2
pypiwin32==223
pyppeteer==2.0.0
PyPrind==2.11.3
pyquery==2.0.0
pyrsistent==0.18.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python-git==2018.2.1
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywin32-ctypes==0.2.3
pywinpty==2.0.8
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq==24.0.1
ratelim==0.1.6
regex==2022.3.15
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
s3transfer==0.12.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
slacker==0.9.65
slacker-log-handler==1.8.0
smart-open==5.2.1
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sounddevice==0.4.5
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SpeechRecognition==3.8.1
Sphinx==5.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.3
stack-data==0.5.1
swing==0.0.4
sympy==1.13.2
tabulate==0.8.10
tenacity==8.1.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.15.0
thinc==8.0.15
threadpoolctl==3.0.0
tinycss2==1.1.1
tkcalendar==1.6.1
tkintermapview==1.11
tkVideo==0.1
tomli==2.0.1
torch==2.4.1
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
typer==0.4.1
typing_extensions==4.12.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
w3lib==2.1.2
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==3.1.3
wget==3.2
wikipedia==1.4.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
xyzservices==2024.4.0
zc.lockfile==3.0.post1
zipp==3.18.2
zope.event==5.0
zope.interface==7.2
